make lowlim for r0 into a double instead of an integer.

+++ analysisrotated.R
@@ -56,7 +56,7 @@ option_list <- list(
     make_option(c("--lowlim"), type = "integer", default = 1,
     help = "points lower than this are not used
             for determining xi [default %default]"),
-    make_option(c("--lowlimpot"), type = "integer", default = -1,
+    make_option(c("--lowlimpot"), type = "double", default = -1,
     help = "points lower than this are not used
             for determining r_0, if negative, the same as lowlim [default %default]"),
     make_option(c("--mindistance"), type = "integer", default = 2,
@@ -152,10 +152,10 @@ if (opt$smearing) {
         Nt, Ns, beta, xi, nape, alpha, opt$bootl, opt$usecov)
 }
 
-endingdofit <- sprintf("Nt%dNs%dbeta%fxi%fbs%domit%dllxi%dllr0%d",
+endingdofit <- sprintf("Nt%dNs%dbeta%fxi%fbs%domit%dllxi%dllr0%.3f",
         Nt, Ns, beta, xi, opt$bootsamples, opt$omit, opt$lowlim, opt$lowlimpot)
 if (opt$smearing) {
-        endingdofit <- sprintf("Nt%dNs%dbeta%fxi%fnape%dalpha%fbs%domit%dllxi%dllr0%d",
+        endingdofit <- sprintf("Nt%dNs%dbeta%fxi%fnape%dalpha%fbs%domit%dllxi%dllr0%.3f",
         Nt, Ns, beta, xi, nape, alpha, opt$bootsamples, opt$omit, opt$lowlim, opt$lowlimpot)
 }
